(#55) Remove theme
The theme was included in the initial commit, but hadn’t been updated
since. The only changes from the stock theme were to the sidebar
coloring in “Ayu” mode.

Removing the theme hides the extraneous `fn main() {` code added to code
blocks and fixes the playground links for code blocks that can be run on
their own.