Add ignore attribute to most Rust code blocks
This adds the [ignore attribute][attr] to Rust code blocks that cannot
be run on their own. The ignore attribute prevents both the playground
and show link icons from appearing. (The show link icon causes `fn
main() {` type code to be displayed around the block, which is often not
appropriate.)

[attr]: https://rust-lang.github.io/mdBook/cli/test.html